Library

Contents

Library

Contents

Types

Functions

Types

CoupledFields.InputSpace — Type

InputSpace(X, Y, d, lat)

A type to hold the X and Y fields of the Input space. The fields are whitened if d=[d1, d2] is supplied. Area weighting is applied if lat is supplied.

CoupledFields.ModelObj — Type

ModelObj(W, R, A, T, evals, pars, method)

A type to hold statistical model results, such as the matrices W, R, A, T, where R=XW and T=YA.

CoupledFields.KernelParameters — Type

KernelParameters

An abstract type.

All KernelParameters types contain certain parameters which are later passed to internal functions Kf and ∇Kf.

A KernelParameters type is set using e.g. PolynomialKP(X::Matrix{Float64}) or GaussianKP(X::Matrix{Float64}).

CoupledFields.GaussianKP — Type

GaussianKP(X)

For the gaussian kernel.

CoupledFields.PolynomialKP — Type

PolynomialKP(X)

For the polynomial kernel.

Functions

CoupledFields.CVfn — Method

CVfn(parm::Matrix, X::Matrix, Y::Matrix, modelfn::Function, kerneltype::DataType; verbose=true, dcv=2)

Cross-validation function

CoupledFields.Rsq_adj — Method

Rsq_adj(Tx::Array, Ty::Array, df::Int)

Cross-validation metric

CoupledFields.bf — Method

bf(x::Vector, df::Int)

Compute a piecewise linear basis matrix for the vector x.

CoupledFields.cca — Method

cca(v::Array, X::Matrix, Y::Matrix)

Regularized Canonical Correlation Analysis using SVD.

CoupledFields.gKCCA — Method

gKCCA(par::Array, X::Matrix, Y::Matrix, kpars::KernelParameters)

Compute the projection matrices and components for gKCCA.

CoupledFields.gradvecfield — Method

gradvecfield(par::Array, X::Matrix, Y::Matrix, kpars::KernelParameters)

Compute the gradient vector or gradient matrix at each instance of the X and Y fields, by making use of a kernel feature space.

CoupledFields.whiten — Method

whiten(X::Matrix, d::Float64; lat=nothing)

Whiten X.

d (0-1) Percentage variance of components to retain.

lat Latitudinal area-weighting.